% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burritos_cost.R
\name{strawberry_puffs}
\alias{strawberry_puffs}
\title{Calculate total happiness from strawberry puffs}
\usage{
strawberry_puffs(sugar, happy)
}
\arguments{
\item{sugar}{Amount of sugar consumed}

\item{happy}{Utility points for resulting happiness}
}
\value{
returns the max happiness from strawberry puffs
}
\description{
Calculate total happiness from strawberry puffs
}
